requirements.txt
pandas
pydantic
pdfplumber
pymupdf
pyyaml
streamlit
rapidfuzz
tqdm
python-dotenv
openai